fix for guest account

+++ roommmatedock/messageboard/views.py
@@ -98,5 +98,10 @@ def delete_message(request, profile_name, date, sender, subject):
     subject=subject,
         )    
     
-    mes.delete()
-    return redirect('messageboard:index', profile_name=profile_name)
+    if reply_profile.name == 'Tommy\'s intern office':
+        messages.warning(request, 'Guest account messages can\'t be deleted')
+        return re
+        return redirect('messageboard:index', profile_name=profile_name)
+    else:    
+        mes.delete()
+        return redirect('messageboard:index', profile_name=profile_name)
